React2Role: Handle quote characters in channel descriptions
Shlex started raising errors about badly quoted text once some kind of
quote appeared in the text (e.g. ').

Instead of that, we're blindly splitting the line by space now. I'm not
sure if the usage of shlex was intentional and it was solving some other
issues; we're not supplying quoted roles anywhere, so it would pick the
same strings either way.